requirements.txt
argo-workflows==6.3.5
certifi==2022.12.7
hera-workflows==4.3.1
python-dateutil==2.8.2
pytz==2022.7
six==1.16.0
urllib3==1.26.13
